+++ src/pages/todo.tsx
@@ -6,6 +6,7 @@ import {
   useRecoilState,
 } from 'recoil'
 import TodoItem from '../components/todoItem'
+import TodoListStats from '../components/todoStats'
 import { todoListState, todoListFilterState } from '../atoms/todo'
 
 const TodoList = () => {
@@ -92,40 +93,5 @@ const filteredTodoListState = selector({
     }
   },
 })
-
-const todoListStatsState = selector({
-  key: 'todoListStatsState',
-  get: ({ get }) => {
-    const todoList = get(todoListState)
-    const totalNum = todoList.length
-    const totalCompletedNum = todoList.filter((item) => item.isComplete).length
-    const totalUncompletedNum = totalNum - totalCompletedNum
-    const percentCompleted =
-      totalNum === 0 ? 0 : (totalCompletedNum / totalNum) * 100
-
-    return {
-      totalNum,
-      totalCompletedNum,
-      totalUncompletedNum,
-      percentCompleted,
-    }
-  },
-})
-
-const TodoListStats = () => {
-  const { totalNum, totalCompletedNum, totalUncompletedNum, percentCompleted } =
-    useRecoilValue(todoListStatsState)
-
-  const formattedPercentCompleted = Math.round(percentCompleted)
-
-  return (
-    <ul>
-      <li>Total items: {totalNum}</li>
-      <li>Items completed: {totalCompletedNum}</li>
-      <li>Items not completed: {totalUncompletedNum}</li>
-      <li>Percent completed: {formattedPercentCompleted}</li>
-    </ul>
-  )
-}
 
 export default TodoList
